+++ backend/.gitignore
@@ -1,2 +1,2 @@
 #ignore models
-models/
+fernec_api/ia_models/
